refactor(muya): split clipboard/index.ts into focused modules
Break the 1217-line Clipboard class into concern-scoped modules, leaving
index.ts as a thin orchestrator (class + event wiring + public API):

- copyData.ts  — getClipboardData / writeClipboardData + table-selection payload
- cut.ts       — cutSelection + removeBlocks tree surgery + table-structure cut
- paste.ts     — pasteSelection + parsed / literal / html-block paste branches
- pasteImage.ts— image-insert helpers

Also clears the ESLint warnings the old file carried: getClipboardData
(233 lines / complexity 27) and _removeBlocks (complexity 23) are decomposed
into named helpers, pasteHandler (221 lines) is split so it drops its
complexity disable, and the image helpers become module functions (no more
leading-underscore naming warnings). Public API and behavior unchanged.
